# Faster MathJax configuration