--- a/doc/specs/vulkan/config/docbook-xsl/xhtml.xsl
+++ b/doc/specs/vulkan/config/docbook-xsl/xhtml.xsl
@@ -17,14 +17,16 @@
 
 <!-- Add MathJax <script> tags  to document <head> -->
 <xsl:template name="user.head.content">
-    <script type="text/x-mathjax-config">
-        MathJax.Hub.Config({
-            MathML: { extensions: ["content-mathml.js"] },
-            tex2jax: { inlineMath: [['$','$'], ['\\(','\\)']] }
-        });
-    </script>
-    <script type="text/javascript"
-        src="https://cdn.mathjax.org/mathjax/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML">
-    </script>
+  <script type="text/x-mathjax-config">
+  MathJax.Hub.Config({
+    tex2jax: {
+      inlineMath: [ ['$','$'], ['\\(','\\)'] ]
+    },
+    "fast-preview": { disabled: true },
+    "AssistiveMML": { disabled: true }
+    });
+  </script>
+  <script type="text/javascript" src="https://cdn.mathjax.org/mathjax/latest/MathJax.js?config=TeX-AMS_CHTML">
+  </script>
 </xsl:template>
 </xsl:stylesheet>