--- a/doc/specs/vulkan/config/mathjax.js
+++ b/doc/specs/vulkan/config/mathjax.js
@@ -1,8 +1,11 @@
-    <script type="text/x-mathjax-config">
-	MathJax.Hub.Config({
-	    MathML: { extensions: ["content-mathml.js"] },
-	    tex2jax: { inlineMath: [['$','$'], ['\\(','\\)']] }
-	});
-    </script>
-    <script type="text/javascript" src="https://cdn.mathjax.org/mathjax/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML">
-    </script>
+  <script type="text/x-mathjax-config">
+  MathJax.Hub.Config({
+    tex2jax: {
+      inlineMath: [ ['$','$'], ['\\(','\\)'] ]
+    },
+    "fast-preview": { disabled: true },
+    "AssistiveMML": { disabled: true }
+    });
+  </script>
+  <script type="text/javascript" src="https://cdn.mathjax.org/mathjax/latest/MathJax.js?config=TeX-AMS_CHTML">
+  </script>