@krOoze 2016-05-09
BTW I was also experimenting with KaTeX in https://github.com/krOoze/Vulkan-Docs/tree/KaTeX . Seems faster/immediate, but won't render everything, so not a fair comparison and not a direct replacement. Also it doesn't work correctly in the XHTML Transitional schema (only in the html version of the spec)...