@oddhack 2016-05-10
This seems very promising - thanks! Will bring up for evaluation by the Vulkan working group.

I'm kinda hoping there's some way to pre-generate the equation HTML eventually but haven't explored that. We will be moving to HTML5 whenever we can transition to asciidoctor - asciidoc just doesn't support that as an output target.