@TobiasHector 2016-05-10
@oddhack FYI there's a toolchain to generate SVG files from latexmath that has been integrated with AsciiDoctor, which I'm looking into along with the rest of the transition. I was only going to use it for the pdf output, but I can look into doing it for the html outputs too. Then it just means loading up SVGs when the page loads, which should be much quicker.