@krOoze 2016-05-10
> I'm kinda hoping there's some way to pre-generate the equation HTML eventually but haven't explored that.

More than likely there is. I just couldn't stand it anymore and went after the low hanging fruit for the short term.

@TobiasHector Now that you brought it up, how's the migration to AsciiDoctor progressing anyway?