@TobiasHector 2016-05-10
> @TobiasHector Now that you brought it up, how's the migration to AsciiDoctor progressing anyway?

It's getting there. It's going to be a little painful to transition due to some awkward incompatibilities that mean I'll have to touch multiple portions of the spec for single line fixes (mostly equations and tables) - but I can generate documents for most of the targets with 90% of the content there with some quick fixes. Equations are the main missing piece right now. Unfortunately the last 10% is always the hardest bit ;)