@oddhack 2016-05-12
I checked this out on my machine (AMD FX-6300 hexacore, Debian 8, Chromium 50.0). I'm not entirely sure how to measure processing time other than counting from when I hit return in the address bar until the status updates from MathJax in the lower left stop updating, but if that's a good measure, the results actually seem to be worse with this change. About 7-8 seconds for the 1.0 spec with the old MathJax config, about 17 seconds with the new one. I thought maybe there was some caching effect taking place and tried reloading a couple of times, and then opening in a new Chromium user window, but there didn't seem to be much change. Any thoughts on what I'm seeing here? I believe another Khronos person tried on their machine with similar results.