@oddhack 2016-05-12
The main observable difference is after the initial loading phase, I see a bunch of "Typesetting math: ##%" messages, which correspond to the extra time.