@oddhack 2016-05-12
I was unfamiliar with KaTeX before seeing this. It does seem potentially promising. We use the align* environment a lot but I suspect that could be replaced by aligned.

The trickiest part for me was getting math imbedded in asciidoc to flow through properly to both MathJax-in-XHTML and through Docbook to dblatex. Now we are looking at moving to asciidoctor and it's a similar problem, with different details. They have a direct PDF generator but AFAICT the  hoops you have to jump through to hope to deal with math on that path are many and varied. So we would probably end up, at least initially, with asciidoctor->HTML5 and asciidoctor->Docbook->dblatex paths. Unfortunately part of the motivation for asciidoctor was to speed the build times, and since the PDF path would be essentially unchanged, this wouldn't help as much as we'd like.