@ratchetfreak 2016-05-12
Chrome and Firefox both have a profiler tab in their dev tools both have the same keyboard shortcut  to access them (ctrl+shift+I).