@TobiasHector 2016-05-12
> at least initially, with asciidoctor->HTML5 and asciidoctor->Docbook->dblatex paths. Unfortunately part of the motivation for asciidoctor was to speed the build times, and since the PDF path would be essentially unchanged, this wouldn't help as much as we'd like.

We should have 'direct', and more importantly *identical* paths for both, with math flowing through the "mathematical" ruby gem, which would generate svg files from latexmath, which are then included in the html, and rendered into the pdf. Ideally I'd like to ditch the docbook toolchain altogether, and it's looking like that shouldn't be a problem.