@krOoze 2016-05-12
@oddhack I should perhaps disclose my mesurement environment and method.

I did it on Celeron G1610, Windows 10 x64, Firefox 46. Nothing fancy though to ensure science grade measurement - just my regular PC. Should have said so earlier. Good thing, you are testing different browsers and setups.

I loaded the page once (so it would cache or whatnot). Then I measured the time from clicking reload until the bold final (not the thin preview) equations appear. Again nothing fancy, just me clicking stopwatch on my phone. Did that three times, to make sure the times are consistent (they are, +-1 s). I can post the raw data, if interested.