@krOoze 2016-05-12
(Additional) measurements:

| browser    | original [s] | proposed [s] |
|------------|--------------:|--------------:|
| Firefox 46 | 74           | 34           |
| IE 11      | 114          | 28           |
| Chrome 50  | 43           | 25           |