@oddhack 2016-05-12
@TobiasHector AFAICT all the Ruby stuff for math-in-direct-PDF was in an incomplete and badly maintained state; it was not just a matter of 'gem install' to attempt to get it working, there were platform binary components which are going to be difficult to deploy (ala python-lxml, but at least that has prebuilt packages that work), and I eventually gave up on attempts to play with it for the time being.