@oddhack 2016-05-12
@krOoze are you opening up the spec at an anchor that shows some visible math, in order to address the preview vs. final issue? Otherwise I'm not sure how to tell, since the start of the document doesn't have any math (and it would be rendered first anyway, presumably). If you're doing that, what anchor are you testing at?