@krOoze 2016-05-12
@oddhack Yes, usually at ch. 2.8 (it's near the beggining so it won't jerk away). Firefox seems to start at the anchor, other browsers start at the beggining of the page until the script finishes and then jump(so you can wait for that too).

Through browser profiler as @ratchetfreak said might be more pro version of doing it...

It doesn't seem to matter which equation you looking at (they "render" the second version invisible and then just switch it on by css I think - which is relatively immediate).

I think the "loading" includes parsing of that script and downloading is only small part of it. As it happens to be, I do also have a [offline/local MathJax branch](https://github.com/krOoze/Vulkan-Docs/tree/offline_mathjax). Haven't tested it for speed though. Intended it only to make version that does not need internet.