@TobiasHector 2016-05-13
@oddhack Correct, but it's not too painful to setup once you have a decent set of instructions (which I'm working on) - no more painful than setup of a latex toolchain IMO! Anyway if we want to discuss this more let's do it offline - but I am working through all the kinks.