@krOoze 2016-10-08
I am just gonna close this.

Recent update (moving some of the math outside MathJax) improved loading times from maddening to psychologically manageable.

I guess there are also still maybe some changes coming.

Also in the meantime anyone can make these changes by right-click context menu on a MathJax element (Math Settings -> Math Renderer). The browser tends to remember (probably through cookies).